pyproject.toml
[project]
name = "llamaocr"
version = "0.1.0"
description = ""
authors = [
    {name = "Javier Cadena",email = "aj-cadena@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
